requirements.txt
pyyaml>=6.0.1
typing-extensions>=4.5.0
langchain>=0.1.0
langchain-openai>=0.0.2
python-dotenv>=1.0.0
pydantic>=2.0.0 